format

+++ src/main/java/com/cibertec/dsw2/Repository/ConsoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConsoleRepository extends JpaRepository<Console,Long> {
+public interface ConsoleRepository extends JpaRepository<Console,Integer> {
 
 }
